requirements.txt
requests-oauthlib==1.3.0
python-dotenv==0.13.0
pymongo==3.10.1
# numpy==1.18.4
# pandas==1.0.3